Add preliminary JavaScript support to `gpt_bpe`.
