+++ resources/data/gpt2-tokenizer/specials.txt
