+++ resources/embedder/go.mod
@@ -1,0 +1,5 @@
+module embedder
+
+go 1.18
+
+require github.com/siongui/goef v0.0.0-20210610184109-d3b60554c5f3
