+++ resources/mmap_js.go
@@ -1,0 +1,13 @@
+//go:build js
+
+package resources
+
+import (
+	"io"
+	"os"
+)
+
+func readMmap(file *os.File) (*[]byte, error) {
+	contents, err := io.ReadAll(file)
+	return &contents, err
+}
